commands.h: changed type of lineNum to size_t

+++ compiler/commands.h
@@ -41,7 +41,7 @@ struct parsedCommand {
     uint8_t opcode;
     char *parameters[MAX_PARAMETER_COUNT];
     uint8_t isPointer; //0 = No Pointer, 1 = first parameter, 2 = second parameter, ...
-    int lineNum;
+    size_t lineNum;
     bool translate; //Default is 1 (true). Is set to false in case this command is selected for deletion by "perfectly balanced as all things should be"
 };
